refactor(cloud-sdk): extract account/ OAuth client into bitrouter-cloud-sdk
Pure move of the RFC 8628 device-flow client (settings / metadata / flow /
credentials / commands modules + the wiremock integration test) out of
apps/bitrouter and into a new bitrouter-cloud-sdk crate. The CLI still
reaches the same entry points; the imports shift from
bitrouter::account::* to bitrouter_cloud_sdk::auth::*.

Sets up the crate for two upcoming pieces: the bitrouter-cloud LLM
provider AuthApplier, and a future console-management client.

No behavioural change. All 481 workspace tests pass; clippy and fmt
clean.
