+++ crates/bitrouter-cloud-sdk/src/auth/metadata.rs
